+++ src/rogallo/widgets/command_line/open_gemini_uri.py
@@ -7,10 +7,12 @@ from textual.widget import Widget
 ##############################################################################
 # Wasat imports.
 from wasat import GeminiURI, URIError
+from wasat.uri import GEMINI_PREFIX
 
 ##############################################################################
 # Local imports.
 from ...messages import OpenLocation
+from ...preflight import is_likely_capsule, is_likely_schemeless_capsule
 from .base_command import InputCommand
 
 
@@ -34,7 +36,10 @@ class OpenGeminiURICommand(InputCommand):
         try:
             uri = GeminiURI(text)
         except URIError:
-            return False
+            if is_likely_schemeless_capsule(text):
+                uri = GeminiURI(f"{GEMINI_PREFIX}{text}")
+            else:
+                return False
         for_widget.post_message(OpenLocation(uri))
         return True